bme280_xxx.py: Check for the conversion to start.
It may happen that the conversion was not started yet when the
first check for conversion done is done. Then the sensor returns
the previous values. The change adds a short wait until the
conversion is started. The longest wait time is about 5 ms.

Side change: Fix the error message for a wrong oversample argument.

Signed-off-by: robert-hh <robert@hammelrath.com>